+++ frontend/src/components/SignIn/login.vue
@@ -1,9 +1,9 @@
 <template>
 	<div class="login">
 		<br /><br />
-		<div class="tag">I D :</div>
+		<div class="tag">I D )</div>
 		<input type="text" id="id" v-model="user.userid" placeholder="ID를 입력해주세요!" /><br /><br />
-		<div class="tag">P W :</div>
+		<div class="tag">P W )</div>
 		<input type="password" id="password" v-model="user.password" placeholder="PW를 입력해주세요!" /><br /><br />
 		<button v-on:click="login">로그인</button>
 		<p>
@@ -66,12 +66,15 @@ export default {
 
 .tag {
 	position: relative;
-	margin-left: -175px;
+	margin-left: -210px;
+	margin-bottom: 5px;
+	font-size: 15px;
+	color: black;
 }
 
 input {
 	margin: 0px 0;
-	width: 20%;
+	width: 22%;
 	padding: 15px;
 }
 
@@ -87,10 +90,12 @@ button {
 p {
 	margin-top: 40px;
 	font-size: 15px;
+	color: #248657;
 }
 
 p a {
 	text-decoration: underline;
 	cursor: pointer;
+	color: #df0174;
 }
 </style>